further tweaks to finalizing box provisioning. This should fix service not available after a reboot

+++ vm-files/provisioning/install-wordpress.sh
@@ -1,14 +1,18 @@
 #!/bin/bash
 
-# let's save a copy of the wp-cli phar
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+if [[ ! -f "/usr/local/bin/wp" ]]; 
+then
+	echo "Installing WP-CLI..."
+	# let's save a copy of the wp-cli phar
+	curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
 
-# chmod and mv for global usage
-chmod +x wp-cli.phar
-mv wp-cli.phar /usr/local/bin/wp
+	# chmod and mv for global usage
+	chmod +x wp-cli.phar
+	mv wp-cli.phar /usr/local/bin/wp
 
-# does it STILL work?
-wp --info
+	# does it STILL work?
+	wp --info --allow-root
+fi
 
 # Init Script for WordPress and WooCommerce Website
 echo "Setting up WP and WooCommerce E-Commerce Platform base"
@@ -19,7 +23,7 @@ mysql -u root --password=root -e "GRANT ALL PRIVILEGES ON portalrevamp_dev.* TO 
 
 cd /srv/portalrevamp.dev/public_html/
 
-if [ ! -f "wp-config.php" ]
+if [[ ! -f "wp-config.php" ]];
 then
     echo "Downloading Wordpress Core"
     # Download and unzip latest version of WordPress
@@ -29,16 +33,22 @@ then
     # Use WP CLI to install WordPress
     wp core install --url="https://portalrevamp.dev" --title="The Truth About Cancer" --admin_user="admin" --admin_password='ttac2016$$' --admin_email="admin@thetruthaboutcancer.com" --allow-root
     
-    # use wp to change active theme
-    wp theme update storefront --allow-root
-    wp theme activate storefront-child --allow-root
+    # use wp-cli to update/activate theme. 
+    # the "--allow-root" flag must be included.
+    # This provisioning script executes as root
+    # 
+    # wp theme update storefront --allow-root
+    # wp theme activate storefront-child --allow-root
 
     # use wp to activate all appropriate plugins
-    wp plugin delete hello --allow-root
-    wp plugin activate --all --allow-root
+    #wp plugin delete hello --allow-root
+    #wp plugin activate --all --allow-root
+    
+    # tell wordpress that we want to user pretty links
+    wp rewrite structure '/%postname%'
 fi
 
 # The Vagrant site setup script will restart Nginx for us
 
 # Let the user know the good news
-echo "WordPress Stable now installed for bwmc.dev";
+echo "WordPress Stable now installed for portalrevamp.dev";
